flask blueprint for settings
separation of the settings page and sub-pages into a flask blueprint folder and files
