+++ arm/ui/settings/templates/settings/general.html
