+++ arm/ui/settings/templates/settings/ripper.html
