+++ arm/ui/settings/templates/settings/settings.html
